+++ src/Movies.Api/Core/Constants/CachePolicies.cs
@@ -2,7 +2,7 @@
 // Copyright (c) 2025 Junaid Atari, and contributors
 // Repository:https://github.com/blacksmoke26/csharp-webapp
 
-namespace Movies.Api.Core.Caching;
+namespace Movies.Api.Core.Constants;
 
 public static class CachePolicies {
   public static class Movies {
